Fix Station1〜4ではテーブルしないことへの考慮

+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
+from sqlalchemy import inspect
 
-from app.database import SessionLocal
+from app.database import SessionLocal, engine
 from app.models import item_model, list_model
 
 
@@ -8,12 +9,25 @@ from app.models import item_model, list_model
 def db_session():
     db = SessionLocal()
     try:
-        db.query(list_model.ListModel).delete()
-        db.query(item_model.ItemModel).delete()
-        db.commit()
+        _reset_records(db)
         yield db
     finally:
-        db.query(list_model.ListModel).delete()
-        db.query(item_model.ItemModel).delete()
-        db.commit()
+        _reset_records(db)
         db.close()
+
+
+def _reset_records(db) -> None:
+    """テーブルのレコードをリセット."""
+    inspector = inspect(engine)
+    table_names = inspector.get_table_names()
+    is_deleted = False
+
+    if ("todo_lists" in table_names):
+        db.query(list_model.ListModel).delete()
+        is_deleted = True
+    if ("todo_items" in table_names):
+        db.query(item_model.ItemModel).delete()
+        is_deleted = True
+
+    if is_deleted:
+        db.commit()
